Archive/requirements.txt
requests
google-generativeai
streamlit
pandas
python-dotenv
feedparser